Correciones del pull requests
