+++ apps/cartelera-cine/src/app/app.tsx
@@ -7,36 +7,9 @@ import Sidebar from '../components/SideBar';
 import SearchBar from '../components/SearchBar';
 import MovieDetail from '../components/MovieDetail';
 import moviesData from '../movies.json';
+import { movieImages } from '../constants/images';
 import '../styles/App.css';
 
-// Importa las imágenes
-import pelicula1 from '../assets/lord_of_the_rings.jpg';
-import pelicula2 from '../assets/el_hobbit.jpg';
-import pelicula3 from '../assets/van_helsing.jpg';
-import pelicula4 from '../assets/harry_potter.jpg';
-import pelicula5 from '../assets/la_monja.jpg';
-import pelicula6 from '../assets/annabelle.jpg';
-import pelicula7 from '../assets/terrifier.jpg';
-import pelicula8 from '../assets/john_wick.jpg';
-import pelicula9 from '../assets/volver_al_futuro_3.jpg';
-import pelicula10 from '../assets/shrek_2.jpg';
-import pelicula11 from '../assets/piratas_del_caribe.jpg';
-import pelicula12 from '../assets/school_of_rock.jpg';
-
-const images: { [key: number]: string } = {
-  1: pelicula1,
-  2: pelicula2,
-  3: pelicula3,
-  4: pelicula4,
-  5: pelicula5,
-  6: pelicula6,
-  7: pelicula7,
-  8: pelicula8,
-  9: pelicula9,
-  10: pelicula10,
-  11: pelicula11,
-  12: pelicula12,
-};
 
 const App: React.FC = () => {
   const [movies, setMovies] = useState<Movie[]>([]);
@@ -46,7 +19,7 @@ const App: React.FC = () => {
   useEffect(() => {
     const updatedMovies = moviesData.movies.map((movie: Movie) => ({
       ...movie,
-      poster: images[movie.id] || '',
+      poster: movieImages[movie.id] || '',
     }));
     setMovies(updatedMovies);
     setFilteredMovies(updatedMovies);
