+++ apps/cartelera-cine/src/components/Header.tsx
@@ -1,15 +1,17 @@
 import React from 'react';
-import '../styles/Header.css'; 
+import '../styles/Header.css';
 import logo from '../assets/logo.png';
 
-const Header: React.FC = () => (
+const Header: React.FC = () => {
+  const cinename = "CINE";
+   return(
   <header className="banner">
     <div className="banner-content">
       <img src={logo} alt="Cine Logo" className="banner-logo" />
-      <h1 className="banner-title">Cine Amarillo</h1>
-   
+      <h1 className="banner-title">{cinename}</h1>
+
     </div>
   </header>
 );
-
+};
 export default Header;
