+++ apps/cartelera-cine/src/constants/images.ts
@@ -1,0 +1,27 @@
+import pelicula1 from '../assets/lord_of_the_rings.jpg';
+import pelicula2 from '../assets/el_hobbit.jpg';
+import pelicula3 from '../assets/van_helsing.jpg';
+import pelicula4 from '../assets/harry_potter.jpg';
+import pelicula5 from '../assets/la_monja.jpg';
+import pelicula6 from '../assets/annabelle.jpg';
+import pelicula7 from '../assets/terrifier.jpg';
+import pelicula8 from '../assets/john_wick.jpg';
+import pelicula9 from '../assets/volver_al_futuro_3.jpg';
+import pelicula10 from '../assets/shrek_2.jpg';
+import pelicula11 from '../assets/piratas_del_caribe.jpg';
+import pelicula12 from '../assets/school_of_rock.jpg';
+
+export const movieImages: { [key: number]: string } = {
+  1: pelicula1,
+  2: pelicula2,
+  3: pelicula3,
+  4: pelicula4,
+  5: pelicula5,
+  6: pelicula6,
+  7: pelicula7,
+  8: pelicula8,
+  9: pelicula9,
+  10: pelicula10,
+  11: pelicula11,
+  12: pelicula12,
+};
