+++ apps/cartelera-cine/src/core/models/Movie.ts
@@ -1,0 +1,11 @@
+export type Movie = {
+    id: number;
+    title: string;
+    genre: string;
+    schedule: string;
+    type: string;
+    rating: string;
+    duration: number;
+    poster?: string;
+    synopsis: string;
+  };
