+++ apps/cartelera-cine/src/styles/Footer.css
@@ -1,7 +1,7 @@
 .footer {
   background-color: #1a1a1a;
-  color: rgba(255, 255, 255, 0.9); /* Color del texto  */
-  text-align: center; /* Centra el texto */
+  color: rgba(255, 255, 255, 0.9); 
+  text-align: center;
   padding: 20px;
 }
 
@@ -11,5 +11,5 @@
 }
 
 .footer a:hover {
-  text-decoration: underline; /* Subrayado al pasar el cursor */
+  text-decoration: underline; 
 }
