+++ apps/cartelera-cine/src/styles/MovieList.css
@@ -2,16 +2,16 @@
   display: flex;
   flex-wrap: wrap;
   justify-content: center;
-  gap: 40px; /* Aumentar el espacio entre los posters */
+  gap: 40px; 
   padding: 20px;
-  margin-left: 20px; /* Espacio adicional para el contenido */
+  margin-left: 20px; 
 }
 
 .movie-card {
-  flex: 0 1 calc(30% - 40px); /* Permitir más espacio */
-  max-width: 250px; /* Aumentar el ancho máximo del poster */
+  flex: 0 1 calc(30% - 40px); 
+  max-width: 250px; 
   position: relative;
-  border: 1px solid rgb(255, 255, 255); /* Marco blanco */
+  border: 1px solid rgb(255, 255, 255); 
   border-radius: 10px;
   overflow: hidden;
   transition: transform 0.3s, box-shadow 0.3s;
@@ -25,9 +25,9 @@
 
 .movie-poster {
   width: 100%;
-  height: auto; /* Mantener proporción de la imagen */
-  max-height: 350px; /* Aumentar altura máxima para que no se corte */
-  object-fit: cover; /* Mantiene proporción de la imagen */
+  height: auto; 
+  max-height: 350px; 
+  object-fit: cover; 
   border-radius: 10px 10px 0 0;
   opacity: 1;
 }
@@ -37,50 +37,50 @@
   bottom: 0;
   left: 0;
   right: 0;
-  background-color: rgba(0, 0, 0, 0.7); /* Fondo negro semi-transparente */
-  padding: 10px; /* Espacio alrededor del texto */
-  border-radius: 10px 10px 0 0; /* Redondear las esquinas superiores */
+  background-color: rgba(0, 0, 0, 0.7); 
+  padding: 10px; 
+  border-radius: 10px 10px 0 0; 
   text-align: left;
-  color: #fff; /* Color del texto blanco */
-  opacity: 0; /* Comenzar oculto */
-  transition: opacity 0.3s ease, box-shadow 0.3s ease; /* Transición suave */
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.5); /* Sombra ligera */
+  color: #fff; 
+  opacity: 0; 
+  transition: opacity 0.3s ease, box-shadow 0.3s ease;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.5); 
 }
 
 .movie-card:hover .movie-overlay {
-  opacity: 1; /* Mostrar al hacer hover */
-  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.7); /* Sombra más pronunciada con efecto ovalado */
+  opacity: 1; 
+  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.7); 
 }
 
 .movie-title {
-  font-size: 1.2rem; /* Reducir el tamaño del título */
-  color: #fff; /* Color blanco */
+  font-size: 1.2rem; 
+  color: #fff; 
   margin: 0;
-  font-family: 'Poppins', sans-serif; /* Fuente del título */
-  font-weight: 500; /* Menos grueso */
+  font-family: 'Poppins', sans-serif; 
+  font-weight: 500;
 }
 
 .movie-details {
   display: flex;
   justify-content: space-between;
   margin-top: 5px;
-  font-size: 0.8rem; /* Reducir el tamaño del texto para detalles */
+  font-size: 0.8rem; 
   font-family: 'Verdana', sans-serif;
 }
 
 .movie-genre, .movie-duration {
-  color: #fff; /* Texto blanco */
-  text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.7); /* Sombra para dar profundidad */
+  color: #fff; 
+  text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.7); 
 }
 
 @media (max-width: 1200px) {
   .movie-card {
-    flex: 0 1 calc(45% - 40px); /* Dos posters en pantallas medianas */
+    flex: 0 1 calc(45% - 40px); 
   }
 }
 
 @media (max-width: 900px) {
   .movie-card {
-    flex: 0 1 calc(100% - 40px); /* Un poster en pantallas más pequeñas */
+    flex: 0 1 calc(100% - 40px); 
   }
 }
