+++ apps/cartelera-cine/src/styles/SearchBar.css
@@ -4,18 +4,18 @@
   align-items: center;
   gap: 10px;
   margin: 20px 0;
-  max-width: 600px; /* Ajustar el ancho máximo */
-  margin-left: auto; /* Centrar la barra de búsqueda en el espacio disponible */
-  margin-right: auto; /* Asegúrate de que se mantenga centrada */
+  max-width: 600px; 
+  margin-left: auto; 
+  margin-right: auto; 
 }
 
 .search-bar input[type="text"] {
-  padding: 10px; /* Aumentar el padding */
+  padding: 10px; 
   border: 2px solid rgba(238, 236, 236, 0.8);
   box-shadow: 0 0 20px #512c71;
   border-radius: 5px;
   outline: none;
-  flex-grow: 1; /* Permitir que el input ocupe todo el espacio disponible */
+  flex-grow: 1; 
   background-color: rgba(30, 30, 30, 0.8);
   color: #fff;
 }
@@ -27,44 +27,43 @@
   border: none;
   border-radius: 5px;
   cursor: pointer;
-  transition: background-color 0.3s; /* Suavizar el cambio de color */
+  transition: background-color 0.3s;
 }
 
 .search-bar button:hover {
   background-color: rgb(125, 124, 121);
 }
 
-/* Media Queries para hacer la barra de búsqueda responsiva */
 @media (max-width: 768px) {
   .search-bar {
-    flex-direction: column; /* Alinear los elementos verticalmente en pantallas pequeñas */
-    gap: 5px; /* Reducir el espacio entre los elementos */
-    max-width: 100%; /* Asegurarse de que ocupe el 100% del ancho disponible */
-    margin: 10px 0; /* Reducir el margen en pantallas pequeñas */
+    flex-direction: column;
+    gap: 5px; 
+    max-width: 100%; 
+    margin: 10px 0; 
   }
 
   .search-bar input[type="text"] {
-    width: 100%; /* Ajustar el ancho del input para que ocupe el 100% */
-    box-shadow: 0 0 10px #512c71; /* Reducir el tamaño de la sombra */
+    width: 100%; 
+    box-shadow: 0 0 10px #512c71; 
   }
 
   .search-bar button {
-    width: 100%; /* Ajustar el ancho del botón para que ocupe el 100% */
+    width: 100%;
   }
 }
 
 @media (max-width: 480px) {
   .search-bar {
-    gap: 2px; /* Reducir aún más el espacio entre los elementos en pantallas muy pequeñas */
-    margin: 5px 0; /* Reducir el margen en pantallas muy pequeñas */
+    gap: 2px; 
+    margin: 5px 0; 
   }
 
   .search-bar input[type="text"] {
-    padding: 8px; /* Reducir el padding del input */
-    box-shadow: 0 0 5px #512c71; /* Reducir aún más el tamaño de la sombra */
+    padding: 8px;
+    box-shadow: 0 0 5px #512c71; 
   }
 
   .search-bar button {
-    padding: 8px 10px; /* Reducir el padding del botón */
+    padding: 8px 10px; 
   }
 }
