+++ nx.json
@@ -94,8 +94,8 @@
     }
   },
   "defaultBase": "main"
-<<<<<<< Updated upstream
+
 }
-=======
-}
->>>>>>> Stashed changes
+
+
+
